requirements.txt
ccxt==4.2.15
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
schedule==1.2.1 